requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
tensorflow==2.16.1
pillow==10.3.0
numpy==1.26.4
pyyaml==6.0.1
python-dotenv==1.0.1
pydantic==2.7.1
aiofiles==23.2.1
openai==1.30.0
